Control working directory of the new session
The tests are not optimal. In order to deal with a race condition that
the test is ready to continue before the command to write to the console
has even been started.

Right now, the tests merely sleep for a long duration that _should_ be
long enough for the working directory to have been printed on screen.