Create TmuxTarget type
This can be used to run tmux commands that has a "target". A target can
be either a session, windows, or a pane.

Commands that target windows and panes, can receive session ids as targets. Then, the
active window or pane is targeted.

By letting TmuxSession, TmuxWindow, and TmuxPane embed that type it
simplifies running commands targeting any of those.