Fix how the output stream is processed
The output events from tmux control mode can easily be split in the
middle of a line; as well as the control chars to start a new line can
be split.

This processes the output events and generates a new stream of actual
written lines; applicable for listening to apps that don't move the
cursor around.