Refactor how pane data is
Make the general format output parsing more generic, and reduce duplication.

The TDD appraoch is to start with a working solution; not a clean
solution, and refactor in the green.

The first solution to reading information about the tmux pane dimensions
used a lot of code duplication. This removes a lot of the duplication;
and increases the usefulness of the formatted output parser; at least in
signature. Adding support for more arguments still requires
modifications to the regexp; or possibly a recursive parser.